# openapi.yaml
...
openapi: 3.0.0
info:
  title: service-workbench-on-aws
  description: API collection of service-workbench-on-aws
  contact:
    email: aws-chamonix-dev@amazon.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: v1.4.3
...
paths:
...
  /api/users:
    get:
      tags:
        - admins
        - researchers
        - guests
        - internal-guests
        - User
      summary: List all users
      operationId: ListUsers
      description: |
        List all users
      responses:
        '200':
          description: Array of registered users
    post:
      tags:
        - admins
        - User
      summary: Create a new User
      operationId: CreateUser
      description: |
        Create a new User
      requestBody:
        description: User information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: User added
...
  /api/users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UserId'
    put:
      tags:
        - admins
        - User
      summary: Update user
      operationId: UpdateUser
      requestBody:
        description: User information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      description: |
        Update a user
      responses:
        '200':
          description: User updated
  /api/user:
    get:
      tags:
        - admins
        - researchers
        - guests
        - internal guests
        - User
      summary: Get user information of current user
      operationId: getCurrentUser
      description: |
        Get current user information
      responses:
        '200':
          description: user information
    put:
      tags:
        - admins
        - researchers
        - guests
        - internal guests
        - User
      summary: Update user information of current user
      operationId: updateCurrentUser
      description: |
        Update current user information
      responses:
        '200':
          description: updated user information
      requestBody:
        description: fields to update in user information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCurrentUserInfo'
...
  /api/template/{templateId}:
    get:
      tags:
        - admins
        - guests
        - internal-guests
        - researchers
        - CfnTemplate
      summary: Fetch S3 pre-signed URL which points to the requested CloudFormation template
      operationId: GetCfnTemplateById
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/TemplateId'
      description: |
        Fetch S3 pre-signed URL which points to the requested CloudFormation template
      responses:
        '200':
          description: Pre-signed URL pointing to requested CloudFormation template
...
  /api/workflows/drafts:
    get:
      tags:
        - admins
        - Workflows
      summary: List all workflow drafts
      operationId: getWorkflowDrafts
      responses:
        '200':
          description: 'Return array of workflow drafts'
    post:
      tags:
        - admins
        - Workflows
      summary: Create a draft workflow
      operationId: createDraftWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDraftWorkflow'
      responses:
        '200':
          description: 'Return created draft workflow'
  /api/workflows/drafts/{workflowId}:
    put:
      tags:
        - admins
        - Workflows
      summary: Edit draft workflow
      operationId: editDraftWorkflow
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/WorkflowId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditDraftWorkflow'
      responses:
        '200':
          description: 'Return edited draft workflow'
    delete:
      tags:
        - admins
        - Workflows
      summary: Delete draft workflow
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/WorkflowId'
      responses:
        '200':
          description: 'Return empty object'
  /api/workflows/drafts/publish:
    post:
      tags:
        - admins
        - Workflows
      summary: Publish draft workflow
      operationId: publishDraftWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishDraftWorkflow'
      responses:
        '200':
          description: 'Return published draft workflow'
...
  /api/workflows/{workflowId}:
    get:
      tags:
        - admins
        - Workflows
      summary: Get requested workflow
      operationId: getWorkflowById
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/WorkflowId'
      responses:
        '200':
          description: 'Return requested workflow'
  /api/workflows/{workflowId}/latest:
    get:
      tags:
        - admins
        - Workflows
      summary: Get latest version of requested workflow
      operationId: getLatestWorkflowById
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/WorkflowId'
      responses:
        '200':
          description: 'Return latest version of requested workflow'
  /api/workflows/{workflowId}/v/{versionId}:
    get:
      tags:
        - admins
        - Workflows
      summary: Get requested workflow by version ID
      operationId: getWorkflowByVersionId
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/WorkflowId'
        - name: versionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/VersionId'
      responses:
        '200':
          description: 'Return requested workflow'
  /api/workflows/{workflowId}/v:
    post:
      tags:
        - admins
        - Workflows
      summary: Create a workflow with given workflowId
      operationId: createWorkflow
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/WorkflowId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflow'
      responses:
        '200':
          description: 'Return created workflow'
  /api/workflow-templates/drafts:
    get:
      tags:
        - admins
        - WorkflowTemplates
      summary: Get draft workflow templates
      operationId: getDraftWorkflowTemplates
      responses:
        '200':
          description: 'Return array of draft workflow templates'
    post:
      tags:
        - admins
        - WorkflowTemplates
      summary: Created draft workflow template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDraftWorkflowTemplate'
      operationId: createDraftWorkflowTemplate
      responses:
        '200':
          description: 'Return created draft workflow template'
  /api/workflow-templates/drafts/{workflowTemplateId}:
    put:
      tags:
        - admins
        - WorkflowTemplates
      summary: Edit draft workflow template
      parameters:
        - name: workflowTemplateId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/WorkflowTemplateId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishDraftWorkflowTemplate'
      operationId: editDraftWorkflowTemplate
      responses:
        '200':
          description: 'Return edited draft workflow template'
    delete:
      tags:
        - admins
        - WorkflowTemplates
      summary: Delete draft workflow template
      parameters:
        - name: workflowTemplateId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/WorkflowTemplateId'
      operationId: deleteDraftWorkflowTemplate
      responses:
        '200':
          description: 'Return empty object'
  /api/workflow-templates/drafts/publish:
    post:
      tags:
        - admins
        - WorkflowTemplates
      summary: Publish draft workflow template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishDraftWorkflowTemplate'
      operationId: publishDraftWorkflowTemplate
      responses:
        '200':
          description: 'Return created workflow template'
...
  /api/workflow-templates/{workflowTemplateId}:
    get:
      tags:
        - admins
        - WorkflowTemplates
      summary: Get all versions of requested workflow template
      parameters:
        - name: workflowTemplateId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/WorkflowTemplateId'
      operationId: getWorkflowTemplate
      responses:
        '200':
          description: 'Return array of all versions of requested workflow template'
  /api/workflow-templates/{workflowTemplateId}/latest:
    get:
      tags:
        - admins
        - WorkflowTemplates
      summary: Get latest version of requested workflow template
      parameters:
        - name: workflowTemplateId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/WorkflowTemplateId'
      operationId: getLatestWorkflowTemplateById
      responses:
        '200':
          description: 'Return latest version of requested workflow template'
  /api/workflow-templates/{workflowTemplateId}/v/{versionId}:
    get:
      tags:
        - admins
        - WorkflowTemplates
      summary: Get requested workflow template by id and versionId
      parameters:
        - name: workflowTemplateId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/WorkflowTemplateId'
        - name: versionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/VersionId'
      operationId: getWorkflowTemplateByIdAndVersion
      responses:
        '200':
          description: 'Return requested workflow template'
  /api/workflow-templates/{workflowTemplateId}/v:
    post:
      tags:
        - admins
        - WorkflowTemplates
      summary: Created new workflow template
      parameters:
        - name: workflowTemplateId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/WorkflowTemplateId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowTemplate'
      operationId: createWorkflowTemplate
      responses:
        '200':
          description: 'Return created workflow template'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TemplateId:
      type: string
      example: 'sagemaker.cfn.yml'
...
    VersionId:
      type: number
      example: 2
...
    WorkflowId:
      type: string
      minLength: 1
      maxLength: 100
      pattern: "^[A-Za-z0-9-_]+$"
      example: 'wf-create-environment'
...
    WorkflowTemplateId:
      type: string
      minLength: 1
      maxLength: 100
      pattern: "^[A-Za-z0-9-_]+$"
      example: 'wt-empty'
    CreateWorkflow:
      type: object
      properties:
        workflowTemplateId:
          $ref: '#/components/schemas/WorkflowTemplateId'
        selectedSteps:
          $ref: '#/components/schemas/SelectedStep'
        desc:
          type: string
        workflowTemplateVer:
          type: number
        v:
          type: number
        id:
          type: string
          minLength: 1
          maxLength: 100
          pattern: "^[A-Za-z0-9-_]+$"
        title:
          type: string
      example: {
        "workflowTemplateId": "wt-empty",
        "selectedSteps": [ ],
        "desc": "Start a Service Catalog SageMaker Jupyter Notebook.\n",
        "workflowTemplateVer": 1,
        "v": 1,
        "title": "Start a Service Catalog SageMaker Notebook",
        "id": "foo-5-workflow",
      }
    CreateDraftWorkflow:
      type: object
      properties:
        selectedSteps:
          $ref: '#/components/schemas/SelectedSteps'
        desc:
          type: string
        workflowTemplateVer:
          type: number
        v:
          type: number
        title:
          type: string
        templateId:
          type: string
          minLength: 1
          maxLength: 100
          pattern: "^[A-Za-z0-9-_]+$"
        workflowId:
          type: string
          minLength: 1
          maxLength: 100
          pattern: "^[A-Za-z0-9-_]+$"
      example: {
        "workflowId": "foo-1-draft-workflow",
        "selectedSteps": [ ],
        "desc": "Start a Service Catalog SageMaker Jupyter Notebook.\n",
        "workflowTemplateVer": 1,
        "v": 1,
        "title": "Start a Service Catalog SageMaker Notebook",
        "templateId": "wt-empty",
      }
    EditDraftWorkflow:
      type: object
      properties:
        rev:
          type: number
        workflowId:
          type: string
        templateId:
          type: string
        id:
          type: string
        workflow:
          type: object
      example: {
        "rev": 6,
        "workflowId": "wf-foo-1-draft-workflow",
        "templateId": "wt-empty",
        "updatedBy": "u-uodoD9RlS9IAOfwpFyqwQ",
        "id": "u-uodoD9RlS9IAOfwpFyqwQ_wf-foo-1-draft-workflow_0",
        "workflow": {
          "id": "wf-foo-1-draft-workflow",
          "selectedSteps": [ ],
          "desc": "An empty workflow so that you have full control of the workflow.\n",
          "workflowTemplateVer": 1,
          "v": 1,
          "title": "Empty Workflow",
          "workflowTemplateId": "wt-empty",
          "updatedAt": "2021-01-28T22:04:16.341Z"
        },
      }
    PublishDraftWorkflow:
      type: object
      properties:
        rev:
          type: number
        workflow:
          type: object
        workflowId:
          type: string
        templateVer:
          type: number
        templateId:
          type: string
        workflowVer:
          type: number
        updatedBy:
          type: string
        id:
          type: string
        createdBy:
          type: string
      example: {
        "rev": 0,
        "workflow": {
          "rev": 0,
          "updatedBy": "u-uodoD9RlS9IAOfwpFyqwQ",
          "selectedSteps": [ ],
          "builtin": false,
          "title": "Empty Workflow",
          "createdAt": "2021-01-28T22:31:03.542Z",
          "workflowTemplateId": "wt-empty",
          "createdBy": "u-uodoD9RlS9IAOfwpFyqwQ",
          "v": 1,
          "id": "wf-foo-5-draft-workflow",
          "workflowTemplateVer": 1,
          "desc": "An empty workflow so that you have full control of the workflow.\n",
          "updatedAt": "2021-01-28T22:31:03.542Z"
        },
        "workflowId": "wf-foo-5-draft-workflow",
        "templateVer": 1,
        "templateId": "wt-empty",
        "workflowVer": 0,
        "updatedBy": "u-uodoD9RlS9IAOfwpFyqwQ",
        "id": "u-uodoD9RlS9IAOfwpFyqwQ_wf-foo-5-draft-workflow_0",
        "createdBy": "u-uodoD9RlS9IAOfwpFyqwQ"
      }
    CreateWorkflowTemplate:
      type: object
      properties:
        desc:
          type: string
        id:
          type: string
        v:
          type: number
        selectedSteps:
          $ref: '#/components/schemas/SelectedSteps'
        title:
          type: string
        propsOverrideOption:
          type: object
      example: {
        "desc": "Start a Service Catalog SageMaker Jupyter Notebook.\n",
        "id": "foo-1-workflow",
        "v": 1,
        "selectedSteps": [ ],
        "title": "Start a Service Catalog SageMaker Notebook",
        "propsOverrideOption": { }
      }
    CreateDraftWorkflowTemplate:
      type: object
      properties:
        desc:
          type: string
        templateId:
          type: string
        title:
          type: string
      example: {
        "desc": "An empty workflow so that you have full control of the workflow.\n",
        "templateId": "wt-draft-foo5",
        "title": "Empty Workflow"
      }
    PublishDraftWorkflowTemplate:
      type: object
      properties:
        id:
          type: string
        template:
          type: object
        templateId:
          type: string
      example: {
        "id": "u-uodoD9RlS9IAOfwpFyqwQ_wt-draft-foo-7_0",
        "template": {
          "id": "wt-draft-foo-7",
          "propsOverrideOption": { },
          "selectedSteps": [ ],
          "title": "Untitled",
          "v": 1
        },
        "templateId": "wt-draft-foo7"
      }
    SelectedStep:
      type: object
      properties:
        stepTemplateId:
          type: string
        configs:
          type: object
        stepTemplateVer:
          type: number
        skippable:
          type: boolean
        id:
          type: string
        title:
          type: string
        desc:
          type: string
      example: {
        "stepTemplateId": "st-start-sagemaker-environment-sc",
        "configs": { },
        "stepTemplateVer": 1,
        "skippable": true,
        "id": "wf-step_1_1517380435823_28",
        "title": "Start SageMaker Environment",
        "desc": "Start a SageMaker Jupyter Notebook.\n"
      }
    SelectedSteps:
      type: array
      items:
        $ref: '#/components/schemas/SelectedStep'
      example: [
        {
          "stepTemplateId": "st-start-sagemaker-environment-sc",
          "configs": { },
          "stepTemplateVer": 1,
          "skippable": true,
          "id": "wf-step_1_1517380435823_28",
          "title": "Start SageMaker Environment",
          "desc": "Start a SageMaker Jupyter Notebook.\n"
        }
      ]
...
    UserId:
      type: string
      example: 'u-C-l14OuIfNk7vL7AqmjvC'
      description: id used to uniquely identify a user
...
    CreateUserRequest:
      required:
        - username
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 300
          pattern: '^[A-Za-z0-9-_.]+$|^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$'
          example: 'testuser'
        usernameInIdp:
          type: string
          minLength: 3
          maxLength: 300
          pattern: '^[A-Za-z0-9-_.]+$|^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$'
          example: 'testidpuser'
        password:
          type: string
          example: 'randompassword'
        authenticationProviderId:
          type: string
          example: 'https://cognito-idp.us-west-2.amazonaws.com/us-west-2_123456789'
        identityProviderName:
          type: string
          example: 'AWS-SSO'
        email:
          type: string
          pattern: '^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$'
          example: 'nobody@example.com'
        firstName:
          type: string
          minLength: 1
          maxLength: 500
          pattern: '^[A-Za-z0-9 .-]+$'
          example: 'Test'
        lastName:
          type: string
          minLength: 1
          maxLength: 500
          pattern: '^[A-Za-z0-9 .-]+$'
          example: 'User'
        userType:
          type: string
          enum:
            - root
          example: 'root'
        isSamlAuthenticatedUser:
          type: boolean
          example: true
        isAdmin:
          type: boolean
          example: true
        status:
          type: string
          enum:
            - active
            - inactive
            - pending
          example: 'active'
        rev:
          type: number
          example: 0
        userRole:
          type: string
          example: 'researcher'
        projectId:
          type: string
          example: 'TestProject'
        isExternalUser:
          type: boolean
          example: true
        encryptedCreds:
          type: string
          example: 'EncryptedCreds123'
        applyReason:
          type: string
          example: 'Research'
    UpdateUserRequest:
      required:
        - uid
        - rev
      type: object
      properties:
        uid:
          $ref: '#/components/schemas/UserId'
        email:
          type: string
          pattern: '^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$'
          example: 'nobody@example.com'
          description: 'Email address of the user'
        firstName:
          type: string
          minLength: 1
          maxLength: 500
          pattern: '^[A-Za-z0-9 .-]+$'
          example: 'Test'
          description: 'First name of the user'
        lastName:
          type: string
          minLength: 1
          maxLength: 500
          pattern: '^[A-Za-z0-9 .-]+$'
          example: 'User'
          description: 'Last name of the user'
        userType:
          type: string
          enum:
            - root
          example: 'root'
          description: 'Specify if user is root'
        isSamlAuthenticatedUser:
          type: boolean
          example: true
          description: 'If using Cognito, is the user authenticated using SAML'
        isAdmin:
          type: boolean
          example: true
          description: 'True if user is admin'
        status:
          type: string
          enum:
            - active
            - inactive
            - pending
          example: 'active'
          description: 'Status of the user. Only active users are allowed access'
        rev:
          type: number
          example: 1
          description: 'Revision specifying the updates the record has gone through'
        userRole:
          type: string
          example: 'researcher'
        projectId:
          type: array
          items:
            type: string
          example: ['project1','project2']
        isExternalUser:
          type: boolean
          example: true
        encryptedCreds:
          type: string
          example: 'EncryptedCreds123'
        applyReason:
          type: string
          example: 'Research'
...
    UpdateCurrentUserInfo:
      type: object
      additionalProperties: false
      properties:
        uid:
          $ref: '#/components/schemas/UserId'
        email:
          type: string
          pattern: '^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$'
          example: 'nobody@example.com'
          description: 'Email address of the user'
        firstName:
          type: string
          maxLength: 500
          pattern: '^[A-Za-z0-9 .-]+$'
          example: 'Test'
          description: 'First name of the user'
        lastName:
          type: string
          maxLength: 500
          pattern: '^[A-Za-z0-9 .-]+$'
          example: 'Test'
          description: 'First name of the user'
        userType:
          type: string
          enum:
            - root
          example: 'root'
          description: 'Specify if user is root'
        isSamlAuthenticatedUser:
          type: boolean
          example: true
          description: 'If using Cognito, is the user authenticated using SAML'
        isAdmin:
          type: boolean
          example: true
          description: 'True if user is admin'
        status:
          type: string
          enum:
            - active
            - inactive
            - pending
          example: 'active'
          description: 'Status of the user. Only active users are allowed access'
        rev:
          type: number
          example: 1
          description: 'Revision specifying the updates the record has gone through'
      required:
        - uid
        - rev
...
security:
  - bearerAuth: []